Refactor. Enhance TxnExecutor
